requirements.txt
isbnlib
requests
lxml
requests-cache
pyyaml
